Address cubic review: frame coalescing, wire hardening, shared example helpers
Inbound stream data is now fed to the reassembly buffer in bounded slices with complete frames drained between top-ups: transport chunking that coalesces several legal RPCs past one frame's cap is no longer misread as a protocol violation (P1), and an oversized chunk can no longer force an unbounded copy before the cap check (P2). Protobuf tags with field number zero are rejected like upstream decoders do instead of being silently skipped.

The examples grow a shared minip2p-example-common crate (key persistence, NAT-event rendering, address shaping) used by both peer and chat -- and the extracted key writer now creates the secret file 0600-from-birth via create_new, fixing the umask window and the check-then-write race that both copies had. The chat e2e captures child stderr into its failure output, the chat loop bounds stdin lines per tick so a piped flood cannot starve the network pump, join-target parsing gets table-driven tests, and CI drops two check steps that were strict subsets of existing clippy/test steps.

Deliberately unchanged: sign_bytes keeps every decoded topic (duplicate-field-4 messages are hand-crafted only; the divergence fails toward rejection and is documented at the site).
