+++ .github/workflows/ci.yml
@@ -29,8 +29,6 @@ jobs:
       - run: cargo clippy --workspace --all-targets -- -D warnings
       - run: cargo clippy -p minip2p --features nat --all-targets -- -D warnings
       - run: cargo clippy -p minip2p --features nat,pubsub --all-targets -- -D warnings
-      - run: cargo check -p minip2p
-      - run: cargo check -p minip2p --features pubsub
       - run: cargo clippy --manifest-path fuzz/Cargo.toml --all-targets -- -D warnings
       - run: cargo check --no-default-features --target thumbv7em-none-eabi -p minip2p-core -p minip2p-identity -p minip2p-transport -p minip2p-tls -p minip2p-identify -p minip2p-multistream-select -p minip2p-ping -p minip2p-pubsub -p minip2p-relay -p minip2p-autonat -p minip2p-dcutr -p minip2p-swarm -p minip2p-nat
       - run: cargo test --workspace
