+++ examples/peer/src/main.rs
@@ -9,7 +9,6 @@
 
 mod cli;
 mod modes;
-mod runtime;
 
 use cli::{CliError, Mode};
